edenapi: rename Fetch to Response
Summary: The `Fetch<T>` type has basically turned into the canonical type EdenAPI for all EdenAPI responses. Originally, this type was merely an implementation detail (essentially just a named tuple returned by the `fetch()` method, hence the name), but given its prominence in the API, the name is confusing. As we add more functionality and usage to this type, it makes sense to give it a more suitable name.

Differential Revision: D30730573

fbshipit-source-id: 7acd2a86b55bdfc186bd9110f6a99333df9d29d3