filenodes: store repo_id and update to new futures
Summary:
Update the `Filenodes` trait so that it doesn't require the repository id to be
passed in every method invocation.  In practice a filenodes instance can only
be used for a single repo, so it is safer for the implementation to store the
repository id.

At the same time, update the trait to use new futures and async-trait.

Reviewed By: yancouto

Differential Revision: D30729630

fbshipit-source-id: a1f80a299d9b0a99ddb267d1f7093f27cf21f1af