blame/deleted_files_manifest: make derivation not depend on BlobRepo
Summary:
Make the derivation process for blame and deleted files manifest not depend
on `BlobRepo`, but instead use the repo attribute (`RepoBlobstore`) directly.
This will allow us to migrate to using `DerivedDataManager` in preparation
for removing `BlobRepo` from derivation entirely.

A `BlobRepo` reference is still needed at the moment for derivation of the
unodes that these depend on.  That will be removed when `DerivedDataManager`
takes over co-ordination of derivation.

Reviewed By: yancouto

Differential Revision: D30729628

fbshipit-source-id: 4504abbe63c9bf036d69cb4341c75b13061fae18