mononoke: add backfill derived data logging
Summary:
We have two mode of deriving data: the "normal" way and using backfilling.
Backfilling is different from "normal" mode in that it derives a few commits at
once, and saves them all to blobstore at once.

Backfilling mode seemed to have helped us when we need to derive a lot of data
(e.g. backfill the whole repo). But

a) We don't know how much it helps, and we don't know if it depends on the repo
b) We don't know if it helps when we derive data for newly landed commits (i.e.
we use "backfill" mode in derived data tailer to derive data for latest public
commits)

So this diff adds a bit of logging to a separate scuba table so that we can get
an idea about things like:
1) How long does it take to derive a stack of commits?
2) Where do we spend most of the time (e.g. deriving, saving the blobs, saving
the mapping).

Reviewed By: mitrandir77

Differential Revision: D30805504

fbshipit-source-id: d82c905cafa87459990d74769a0dddcc91fac174