add logging of client information
Summary:
It can sometimes be difficult to work out from the logging which commit cloud
requests came from which client repo.  Previously you could often infer it from
the client identities, however if the request is proxied, the originating hostname can be
lost, and it still doesn't handle the case where the host contains multiple
repos.

This diff adds a new `ClientInfo` struct, which is included by the client
on every `get_references` and `update_references` request.  This is logged
by the service, allowing us to correlate which client it came from, and what
workspace version the client had at that time.

Reviewed By: StanislavGlebik

Differential Revision: D30697839

fbshipit-source-id: 8fe2e03f0be2f443f8ae1814f083c04ba5d1805e