Restore to correct parent and deleted files
Summary:
This adds basic logic for `snapshot restore` command.
- It updates to the parent of the snapshot
- It loads the snapshot changes

For now I did not do changes/tracked changes, as it will need to download the file contents, which will need a new edenapi endpoint, so I'll leave it for a future diff. It just restores your deleted files for now.

Reviewed By: StanislavGlebik

Differential Revision: D30543507

fbshipit-source-id: 080588ceff0ecd595ce739044f0d4118fb8e1a3f