windows: invalidate negative path cache during start
Summary:
In certain situations, users may cause EdenFS to falsely return a path not exist result while the path is available. Windows will cache that and causing subsequent access to that file to automatically return a file not exist error.

We currently only invalidate this negative cache during checkout and rebooting the machine as the cache is even kept during EdenFS restarts. In this diff, we starts to invalidate the negative path cache at startup so if the user ever had issues an `eden restart` would be sufficient to fix.

Reviewed By: xavierd

Differential Revision: D30814059

fbshipit-source-id: 53283f471702762b2eed0c5d0f6a9cc49f4db739