segmented_changelog: add trait method to test if segmented_changelog is disabled or not
Summary:
In a future diff, we want an endpoint to test if segmented changelog is
supported for a repo without doing any real computation using segmented
chagnelog. This would be useful for the client to decide whether it can
use segmented changelog clone or not, instead of relying on fragile
per-repo configuration.

Reviewed By: farnz

Differential Revision: D30825920

fbshipit-source-id: 16dc5bf762da2d2b9cd808c129e1830285023f3d