scmstore: implement refresh as flush, and forward calls to contentstore when available
Summary: Refresh doesn't have any special meaning outside of packfiles, but in some contexts (like BackingStore) is used to trigger a flush. Previously I'd implemented refresh as a no-op. With this change, refresh will trigger a flush, and if a contentstore fallback is available will also forward the refresh call to contentstore for packfile-specific refresh behavior.

Reviewed By: andll

Differential Revision: D30851223

fbshipit-source-id: 893b4256fd8edc0f60612e61c177d885252cb85b