Fix untracked deletion files being treated as deletion
Summary:
This was a very well-hidden bug that I failed to notice in the integration tests.

Turns out the in serde enum unit variants are serialized to strings, while struct variants to a dictionary.

My code assumed it was always a dictionary. And because python is python, it worked, as the `in` operator works for strings and dicts. But `"Deletion" in fc` for strings means if Deletion is a substring of `fc`, which is also true for `UntrackedDeletion`, thus all untracked deletions were treated like normal deletions.

Reviewed By: StanislavGlebik

Differential Revision: D30868534

fbshipit-source-id: e574de6493bcd8e8d42d6e22da4dc482d083f22d