Basic info command
Summary:
`hg snapshot info` command will be used to get information about the snapshot (similar to `hg show` for commits)

It's still not easy to do this, as we want to have derived data for snapshots, which is still unimplemented.

For now, this makes the command only check if the snapshot exists or not. In the future more functionality will be added (and likely the edenapi endpoint we query will be different).

Reviewed By: liubov-dmitrieva

Differential Revision: D30900088

fbshipit-source-id: 4dc6915d74694a03496c756f03bc073d1a0819f2