http-client: add on_first_activity event
Summary: Add a new `first_activity` event to the `Response` event listeners that fires the first time we detect nonzero progress for either uploading or downloading. This is useful for situations where requests are queued and we want to be notified when the request becomes active (e.g., to register progress bars).

Reviewed By: DurhamG

Differential Revision: D30914242

fbshipit-source-id: 83445724ed81e77ac25954b644e6bbafcbe5cadb