status: remove serial manifest read
Summary:
This line was iterating over a list of files and looking in the
manifest for each one. This results in serial manifest reads which can result in
serial network requests.

Let's instead use manifest.matches() to test them all at once via the underlying
BFS, which does bulk fetching.

Differential Revision: D30938359

fbshipit-source-id: 1af7d417288b82efdd537a4afeaf93c1b55eaf49