backfill_derived_data: fix tests to ensure batch invariants
Summary:
The `backfill_batch_dangerous` method requires that the caller ensures
that all dependencies of the batch have been derived, otherwise errors,
such as mappings being written out before the things they map to, can
occur.

When the derived data manager takes over batch derivation, it will enforce this
requirement, so that it is no longer dangerous.  However, The backfiller tests
were not ensuring the invariant, so the tests will fail with the new derivation
implementation.

Fix the tests by ensuring the parent commits are always derived before a
batch is started.  The test is also extended to expose the failure mode
of accidentally deriving batch parents.  This will be fixed in the next
commit.

Reviewed By: yancouto

Differential Revision: D30959132

fbshipit-source-id: 8489a5d0b375692a903854294e3810846c9e13de