derived_data_utils: extract find_oldest_underived
Summary:
The `find_oldest_underived` method of `DerivedUtils` is used outside tests by
exactly one client (the backfiller in tailing mode).  Simplify the
`DerivedUtils` trait by extracting this method from the trait, and replacing
with a more general one that will be easier to implement in terms of the
derived data manager.

Reviewed By: yancouto

Differential Revision: D30944567

fbshipit-source-id: a1d408e091d145297241a5eebc02a87155bc3765