filenodes: make derivation not depend on BlobRepo
Summary:
Make the derivation process for mercurial filenodes not depend on `BlobRepo`.
Instead, use the repo attributes (`RepoBlobstore` and `Filenodes`) directly.
This will allow us to migrate to using `DerivedDataManager` in preparation
for removing `BlobRepo` from derivation entirely.

The existing use of `changesets` for determining the commit's parents is
changed to use the parents from the bonsai changeset.  For normal derivation,
the bonsai changeset is already loaded, so this saves a database round-trip.
For batch derivation we currently need to load the changeset, but it should
be in cache anyway, as other derived data types will also have loaded it.

We still need to keep a `BlobRepo` reference at the moment.  This is because
filenodes depend on the mercurial derived data.  The recursive derivation is
hidden in the call to `repo.get_hg_from_bonsai_changeset`.  When derivation
is migrated to the derived data manager, we can replace this will a direct
derivation.

Reviewed By: StanislavGlebik

Differential Revision: D30765254

fbshipit-source-id: 20cc17c2eb611544869e5f1c15d858663cd60fd1