repo: separate peer and repo setup
Summary:
Previously repo and peer instantiation were in one unified path. This
allowed treating repo's and peers somewhat interchangably. We're moving to a
world where peers and local repos are quite different, so let's separate these
two paths.

This will be useful in the next diff where we remote the file peer, but want to
keep the ability to instantiate local file non-peer repos.

Reviewed By: quark-zju

Differential Revision: D30975887

fbshipit-source-id: 5e676b522c7cfdd5449aeb6a750947dcb023183f