mononoke: move commit discovery and backfill options initialization
Summary:
I'd like to reuse them in the next diff, so let's refactor it a bit.

Note - in D30837581 (https://github.com/facebookexperimental/eden/commit/315a8b311de1e112048036e516569f51c322a246) markbt suggested a good idea for refactoring
backfill_derived_data. I liked the idea, but when I tried to approach this
refactoring it turned out to be tricky to do so (FWIW, it might be easier to
rewrite everything from scratch). So for now I did the smallest possible
refactoring that's needed to add validation in the next diff, but this small
refactoring can probably be used for a larger refactoring later.

Reviewed By: mitrandir77

Differential Revision: D31115979

fbshipit-source-id: f0b4d70454186a023cd9e12cd645768af1b716e8