mononoke: fix benchmark after migration to derived data manager
Summary:
Derived data manager now doesn't allow deriving a batch of commits if all
ancestors weren't derived yet (and that's a good idea to do this check).

But it started to break benchmark if --batch-size, --backfill and --parallel
options are set, because in the very
beginning of the function we mark all commits as not derived, and when we start
deriving the second batch the first batch is assumed to not be derived, and
this triggers derived data manager check.

Let's instead mark only commits that we are about to derive as not derived, and
clear this check once we are done.

Reviewed By: mitrandir77

Differential Revision: D31140464

fbshipit-source-id: fc74d58dc3c4a3ad70e8e2527f7d6dfc8fde8a9c