mononoke: rename backfill_derived_data benchmark.rs
Summary:
I'd like to use it in the next diffs to add a way to validate that derived data
is the same after rederivation. A lot of the code in `benchmark.rs` is useful
for doing this validation, so let's rename `benchmark.rs` so that it's ok to
use it from two different subcommands.

Reviewed By: mitrandir77

Differential Revision: D31115981

fbshipit-source-id: 86439534d8e49a4022086cb27918b7bcd7befc5c