mononoke: fix edge case of splitting stack when file was replaced with dir
Summary:
We might have a somewhat weird case - a file was replaced with a directory and
then in the next the same file was deleted again (even though this file doesn't exist
anymore). In that case we need to make sure these two commits are in two
different stacks of commits, however previously we weren't doing that. This
diff fixes it.

Reviewed By: markbt

Differential Revision: D31168174

fbshipit-source-id: 4b9986e615ec98b6452ff81b113124d14f236382